chore: (conway) minFeeRefScriptCostPerByte BigInteger -> Integer

+++ components/common/src/main/java/com/bloxbean/cardano/yaci/store/common/domain/ProtocolParams.java
@@ -78,7 +78,7 @@ public class ProtocolParams {
     private BigInteger govActionDeposit; //30
     private BigInteger drepDeposit; //31
     private Integer drepActivity; //32
-    private BigInteger minFeeRefScriptCostPerByte; //33
+    private Integer minFeeRefScriptCostPerByte; //33
     public void merge(ProtocolParams other) {
         if (other.minFeeA != null) {
             this.minFeeA = other.minFeeA;
